requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
line-bot-sdk>=3.0.0
google-generativeai>=0.5.0
yfinance>=0.2.0
python-dotenv>=1.0.0
httpx>=0.27.0
